+++ src/mame/drivers/digel804.cpp
@@ -395,7 +395,7 @@ INPUT_CHANGED_MEMBER( digel804_state::mode_change )
 {
 	if (!newval && !m_keyen_state)
 	{
-		switch ((int)(uintptr_t)param)
+		switch (param)
 		{
 			case MODE_OFF:
 				m_key_mode = m_remote_mode = m_sim_mode = 1;
